move ToggleValue to controllers
