+++ frontend/src/modules/controllers/index.ts
@@ -15,6 +15,7 @@ export { default as ValueSlider } from './ValueSlider';
 export { default as Knob } from './Knob';
 export { default as TabPanel } from './TabPanel';
 export { default as TestTool } from './TestTool';
+export { default as ToogleValue } from './ToggleValue';
 export { default as ValueModal } from './ValueModal';
 export { default as ValueController } from './ValueClicker';
 export { default as AlarmModal } from './AlarmModal';
